requirements.txt
aiogram==3.4.1
openai>=1.30.0
python-dotenv==1.0.1
pydantic==2.5.3
aiohttp==3.9.3
aiosqlite==0.19.0
